Updated tests for formatters
